Various Updates
- Fix DOM element prop warning: Use transientOptions() from this link
https://github.com/emotion-js/emotion/issues/2193#issuecomment-754860641
Example of warning below
- Fix GIFs not having alt warning: add a default if caption doesn't
exist
- Compute custom aspect ratio if image is cropped in Sanity
- Add priority to first 7 rows of images
- Refactor props

Warning: React does not recognize the `backgroundColor` prop on a DOM
element. If you intentionally want it to appear in the DOM as a custom
attribute, spell it as lowercase `backgroundcolor` instead. If you
accidentally passed it from a parent component, remove it from the DOM
element.